feat: Add pre-commit script file (hook)
The idea is to have a way for users to be able to manipulate their repositories
in different ways right before the git-add-git-commit phase. Might be useful in a
lot of different ways: updating arbitrary files that are not directly related to
the package version, but need to have an information about the next version and
being commited before the said version released; or running some chores right
before the commit: including updating docs, dependencies, running scripts, etc.