fix: push only tags when skip-commit and git-push
There's an edge case where if skip-commit is true as well as git-push,
and you follow the instructions for skipping git pull and a commit gets
pushed to the branch while this is running, that it will fail the job.
This is because when pushing the results, it would also push the branch.
Pushing the branch isn't necessary though. This doesn't fix the edge
case though if skip-commit is false but you'd likely want that to fail
then in that case.